+++ R/cli_concatenate.R
@@ -26,3 +26,13 @@ write_summary_by_season <- function(options) {
   summary_by_season <- sum_by_season(data_by_season_and_islets)
   readr::write_csv(summary_by_season, options[["output_path"]])
 }
+
+#' @export
+write_figure_metadata <- function(options) {
+  clean_data <- readr::read_csv(options[["data-path"]], show_col_types = FALSE)
+  figure_metadata_list <- construct_figure_metadata(clean_data, options[["island"]], options[["figure-path"]])
+  readr::write_lines(
+    jsonlite::toJSON(figure_metadata_list, pretty = TRUE, auto_unbox = TRUE),
+    options[["output-path"]]
+  )
+}